second submission

+++ app/Http/Controllers/TasksController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 
 use App\Task;    // 追加
+use App\User;
 
 class TasksController extends Controller
 {
@@ -14,13 +15,21 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {
-         $tasks = Task::all();
-
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+    {  $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+               
+                'tasks' => $tasks,
+            ];
+        }
+        
+        return view('welcome', $data);
+             
     }
+    
 
     /**
      * Show the form for creating a new resource.
@@ -29,7 +38,7 @@ class TasksController extends Controller
      */
     public function create()
     {
-         $task = new Task;
+        $task = new Task;
 
         return view('tasks.create', [
             'task' => $task,
@@ -44,17 +53,20 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-         $this->validate($request, [
+        
+        $this->validate($request, [
             'status' => 'required|max:10',
             'content' => 'required|max:191',
         ]);
-
         
-        $task = new task;
+        
+        $task = new Task;
+        $task->user_id = \Auth::id();
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
-
+        
+        
         return redirect('/');
     }
 
@@ -64,9 +76,9 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    function show($id)
     {
-         $task = Task::find($id);
+        $task = Task::find($id);
 
         return view('tasks.show', [
             'task' => $task,
@@ -79,9 +91,17 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    function edit($id)
     {
         $task = Task::find($id);
+        //タスクを作った人のuserId
+        $taskUserId = $task->user_id;
+        //ログインしている人のuserId
+        $loginUserId = \Auth::id();
+        if($taskUserId != $loginUserId){
+            //トップに飛ばす
+            return redirect('/');
+        }
 
         return view('tasks.edit', [
             'task' => $task,
@@ -95,15 +115,25 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    function update(Request $request, $id)
     {
         $this->validate($request, [
             'status' => 'required|max:10',
             'content' => 'required|max:191',
         ]);
 
-        
         $task = Task::find($id);
+        //タスクを作った人のuserId
+        $taskUserId = $task->user_id;
+        //ログインしている人のuserId
+        $loginUserId = \Auth::id();
+        if($taskUserId != $loginUserId){
+            //トップに飛ばす
+            return redirect('/');
+        }
+            
+        $task = Task::find($id);
+        $task->user_id = \Auth::id();
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
@@ -117,11 +147,22 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    function destroy($id)
     {
+        $task = Task::find($id);
+        //タスクを作った人のuserId
+        $taskUserId = $task->user_id;
+        //ログインしている人のuserId
+        $loginUserId = \Auth::id();
+        if($taskUserId != $loginUserId){
+            //トップに飛ばす
+            return redirect('/');
+        }
+        
         $task = Task::find($id);
         $task->delete();
 
         return redirect('/');
     }
+    
 }
